Coordinates now in f32 not i32

+++ engine/src/game/coordinates.rs
@@ -2,12 +2,12 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Clone, Debug)]
 pub struct Coordinates {
-    x: i32,
-    y: i32,
+    x: f32,
+    y: f32,
 }
 
 impl Coordinates {
-    pub fn new(x: i32, y: i32) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Self { x, y }
     }
 }
